requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.0
python-decouple>=3.8
gunicorn>=20.1.0
whitenoise>=5.3.0
